Fix wrong error message in Checkpoint
When a key was monitored that was not found in history, the resulting
error message would be:

skorch.exceptions.SkorchException: Monitor value 'Key
'valid_loss_best' was not found in history.' cannot be found in history.
Make sure you have validation data if you use validation scores for
checkpointing.

This bugfix now returns the correct error message.