Fix broken tests with new accelerate version
Latest changes in accelerate led to breaking tests. Unfortunately, the
solution is to use a privat method on AcceleratorState but the devs said
that this would be the best solution.

This PR also sets the requirements for accelerate to be v0.11 or higher.
This not only makes our code simpler, but also opens the door to
supporting gradient accumulation with accelerate (which will come in a
separate PR).