Support for gradient accumulation with accelerate (#874)
Support for gradient accumulation with accelerate

The latest accelerate release added gradient accumulation support. This
requires the AccelerateMixin to call the training loop within a context
manager. This is now done. Users can therefore use the gradient
accumulation feature of accelerate.

Furthermore, the learning rate scheduler is now also prepared if it was
used with skorch's LRScheduler callback.

Note:
If users don't use skorch's LRScheduler callback, it cannot be prepared
because there is no reliable way of detecting its use.