Nets now can be loaded even if device=None
In #600, we introduced the option to set device=None, which means that
skorch should not move any data to any device. However, this setting
introduced a bug when trying to load the net, as that code didn't
explicitly deal with device=None. With this PR, the bug is fixed.

Implementation

If device=None, we really have no way of knowing what device to map the
parameters to. The most reasonable thing to do is to use the fallback,
which is 'cpu'.