Fix possible bug when loading a cuda net on CPU
This bug occurs when trying to load a net that was trained on CUDA on a
CPU machine. However, it only occurred when there were CUDA-dependent
attribute set via set_params. This bug is now fixed.

The problem started occurring after PR #751, which introduced storing
parameters set via set_params in the private attribute _kwargs.
Normally, for attributes, we make sure that they can be loaded without
CUDA, but attributes within _kwargs were not checked. Thus, loading
those without CUDA failed. Unfortunately, this was not caught by CI
because CI is not CUDA-enabled.

The solution to this problem is that CUDA-dependent attributes are now
removed from _kwargs during __getstate__ and re-added later during
__setstate__ (a test was introduced for checking that). A little problem
occurring there was that the cuda_dependent_attributes_ are not part of
state. Therefore, I add them to state during __getstate__ and remove
them later in __setstate__.