Remove outdated version checks
Remove version checks that are no longer required because we don't
support the min versions anymore.

I was working on the proposed solution in #887 when I noticed that none
of the version checks we do is actually required anymore. After removing
them, we thus no longer depend on distutils's LooseVersion.

If we ever need to do a version check in the future, I would still go
with the scipy Version class.

Also, in the future, we should ensure to add #TODO comments to all
version checks so that we can easily find them.

One change I introduced here is to require scikit-learn>=1.0.0. I think
this should be fine because it's been out for some time now and it still
supports Python 3.7.

We could go to 0.24 instead and would still not need a version check,
let me know if you prefer that.