Skip GPyTorch tests if torch version too low (#890)
GPyTorch requires relatively high versions of PyTorch. Therefore,
certain tests in the CI matrix may fail for certain GPyTorch versions.
We thus skip the GPyTorch tests in case the PyTorch version is not
sufficiently high.

We don't want to force the PyTorch version itself on the user (as long
as it's sufficiently high) because most users probably don't use
GPyTorch.

To check the versions, as discussed in #888, we use the version checking
code copied from scipy.