Fix false warning if iterator_valid__shuffle=False
Resolves #907

The problem initially occurred because the warning did not check for the
value, just the presence of the key.

Even though there is a test for this, the test didn't detect the error.
This is because during a refactor (#751), the parameter validation was
moved to initialize() from __init__() but the test was not adjusted to
take the change into account.