Don't fail when gpytorch import fails
This is mostly relevant for CI, as most users probably don't have
gpytorch installed. The problem is that some gpytorch versions raise an
AttributeError with some torch versions. Since we try to import gpytorch
in utils.py but only catch ImportError, the AttributeError will be
raised. However, the presence of an incompatible gpytorch version should
not preclude the usage of skorch.

In the gpytorch tests themselves, we already check for compatible
versions and skip if they are not compatible.