Remove top level gpytorch import from utils.py
It is in fact possible to import it only locally with minimal
performance overhead and thus avoid exposing skorch to possible gpytorch
import issues. On top of that, the import can be commented out for now
because it is only needed in case multiclass classification with
gaussian processes will be supported.