Don't fail when gpytorch import fails (#913)
* Don't fail when gpytorch import fails

This is mostly relevant for CI, as most users probably don't have
gpytorch installed. The problem is that some gpytorch versions raise an
AttributeError with some torch versions. Since we try to import gpytorch
in utils.py but only catch ImportError, the AttributeError will be
raised. However, the presence of an incompatible gpytorch version should
not preclude the usage of skorch.

In the gpytorch tests themselves, we already check for compatible
versions and skip if they are not compatible.

* Try fixing gpytorch import error in test

* Remove top level gpytorch import from utils.py

It is in fact possible to import it only locally with minimal
performance overhead and thus avoid exposing skorch to possible gpytorch
import issues. On top of that, the import can be commented out for now
because it is only needed in case multiclass classification with
gaussian processes will be supported.

* Stop warning that GP interface being experimental

It's been there long enough without changes.

* Add a CHANGES.md entry