Update deploy.sh, openssl, use pip (#914)
* Update deploy.sh, openssl, use pip

I ran into this issue when trying to deploy:

https://askubuntu.com/questions/1428181/module-lib-has-no-attribute-x509-v-flag-cb-issuer-check

Therefore, I adjusted the script to update pyopenssl. On top of that, I
was frustrated with the very slow testing cycle, since the script always
creates a new environment from scratch. To accelerate this, I removed
the reliance on the (outdated) environment.yml and used pip for the
requirements.txt, more in line with our CI.

I tested this locally except for the last step (the upload) and it
worked.

* Reviewer comment: better echo message

Co-authored-by: ottonemo <marian.tietz@ottogroup.com>

* Reviewer comment: better comment

Co-authored-by: ottonemo <marian.tietz@ottogroup.com>

Co-authored-by: ottonemo <marian.tietz@ottogroup.com>