Add support for torch.compile (#922)
PyTorch 2.0 has just been released.

It is necessary that we modify skorch if we want to support the feature,
since torch.compile needs to be called on the initialized modules and
initialization happens inside of skorch.

Now users can create a net like NeuralNet(..., compile=True) and all
torch modules should be compiled. By default, compile=False.

To actually see speed improvements, a GPU with a new
architecture (Volta, Ampere) is required.

Implementation

I opted for implementing a single new argument, compile, which can be
set to True. In that case, during initialization of the module and
criterion, the method net.torch_compile will be called on each module
separately and the compiled modules will be stored on the net.

So e.g. the net.module_ will be the compiled module. The original module
would be accessible through net.module_._orig_module:

https://github.com/pytorch/pytorch/blob/b0bd5c4508a8923685965614c2c74e6a8c82f7ba/torch/_dynamo/eval_frame.py#L71

However, given that it's a private attribute, I assume that the compiled
module should just be used like the original module and that shouldn't
cause any trouble.

Furthermore, I wanted users to have the option to easily pass more
arguments to torch.compile, which, as always, works through the use of
dunder names, e.g. NeuralNet(compile=True, compile__mode=...).

If users set compile=True with a torch version that doesn't support
this, an error is raised.

Furthermore, I decided to add a public torch_compile method on
NeuralNet. This method gets the module and the module name, and returns
the compiled module. The idea here is that if users, for instance, want
to skip compiling specific modules, they can do so by checking the
module name.

Since we introduce a new prefix with this PR, testing is a bit more
comprehensive than normal. Please take a look at the TestTorchCompile
test class, which collects all tests related to torch.compile.