Add DistributedHistory for multi-gpu training
Description

(this is more or less copied from the docs)

When training a net in a distributed setting, e.g. when using
torch.nn.parallel.DistributedDataParallel, directly or indirectly with
the help of AccelerateMixin, the default history class should not be
used. This is because each process will have its own history instance
with no syncing happening between processes. Therefore, the information
in the histories can diverge. When steering the training process through
the histories, the resulting differences can cause trouble. When using
early stopping, for instance, one process could receive the signal to
stop but not the other.

DistributedHistory will take care of syncing the distributed batch
information across processes, which will prevent the issue just
described.

This class needs to be initialized with a distributed store provided by
PyTorch (https://pytorch.org/docs/stable/distributed.html#distributed-key-value-store).
I have only tested torch.distributed.TCPStore so far, but
torch.distributed.FileStore should also work. The DistributedHistory
also needs to be initialized with its rank and the world size (number of
processes) so that it has all the required information to perform the
syncing. When using accelerate, that information can be retrieved from
the Accelerator instance.

Comment

Even though the batch information, which is split across processes, is
synced, the epoch information, which is not split, is *not* synced. E.g.
the recorded duration can be different between processes. It is not
quite clear what the "correct" behavior should be here, it would
probably depend on what is done based on this information.

To make it possible to use the new class, I had to change the net
initialization code to not reinitialize the history when it is not None.
Otherwise, calling fit would always overwrite the DistributedHistory
with a normal History object.

Also, unfortunately, TCPStore cannot be pickled. Therefore, I set it
to None when pickling. This is not tragic as long as users pickle the
final model and only load it for inference. If they want to keep on
training, they would need to set the net.history.store manually.