Fix failing tests for GPyTorch v1.10
A couple of changes in GPyTorch v1.10 resulted in our tests failing.
This PR fixes those tests. The actual integration has not changed,
though, only tests.

Be aware that CI will broken until this PR is merged.

Note

The updated tests will now fail for GPyTorch v1.9 and and below.
However, I think it is not worth it to write the tests in a backwards
compatible manner. First of all, the GPyTorch integration does not
appear to be widely used. Second, this only affects the tests, not the
actual implementation, so normal users should have no issues. Developers
would need to upgrade GPyTorch to avoid the error.

Implementation

Initially, we had issues with GPyTorch kernels not being
pickleable/copyable. Therefore, the base testing class tested for those
errors. Now, 2 of the 3 actual test classes (ExactGPRegressor and, since
v1.10, GPRegressor) work. Therefore, I refactored the base test class to
implement the working tests and only the 1 remaining class not working,
GPBinaryClassifier, to implement the tests that raise.

Content-wise, these failing tests have not been changed. The tests for
the working case are now a bit more exact by testing inference results
before and after pickling.

Another issue I encountered was that the default likelihood for
GPBinaryClassifier, BernoulliLikelihood, now no longer has any
arguments. This is annoying because we want to test a few things with
likelihood params, e.g. that params in grid search are passed correctly
to the likelihood (since it is an extra module, this could be a real
source of errors). Therefore, for testing purposes, I created a
likelihood module class that does take an argument (the same as it used
to take before v1.10).