Address #925 (#926)
Changing the `_get_param_names` method to return a list instead of a
generator to fix the exception error message when passing unknown
parameters to `set_params`. Before the error message just included
the generator `repr`-string as the list of possible parameters.
Now the string contains the possible parameter names instead.