Fix #981: replace deprecated `pkg_resources` calls
The `pkg_resources` module is deprecated and mostly superseded
by `importlib`. To find the torch version we don't actually need
to check the metadata, as torch also provides the `__version__`
attribute which can be checked directly.

Therefore, `importlib.metadata.version` is only needed to fetch
the currently installed skorch version as a drop-in replacement
for `pkg_resources.distribution('skorch').version`.