Fix an issue with sklearn
Something somewhere in sklearn changed that now requires the presence of
the classes_ attribute on classifiers when scoring. That attribute is
not always present. Now, more of an effort is made to infer the classes
when a dataset is passed.

Together with the previous commit that enables extraction of y from
TensorDatasets, this makes a test pass that failed otherwise with the
new sklearn version.