MNT: Update CI, add Python 3.11, fix sklearn issues (#990)
* Update checkout and setup-python GH actions
* Add Python 3.11 to CI (only PyTorch 1.13)
* Add pytest-pretty to CI
* Increase columns in GH logs for better readability
* Change check for sklearn error message
* Make data_from_dataset work with TensorDataset

Details

Something somewhere in sklearn changed that now requires the presence of
the classes_ attribute on classifiers when scoring. That attribute is
not always present. Now, more of an effort is made to infer the classes
when a TensorDataset is passed.

An sklearn error message that we checked was changed, resulting in the
corresponding tests to fail. This commit fixes the error message checked
in the test (which was already quite defensive beforehand).