Update docs after accelerate fix was released (#995)
There was an issue with using grid search and other sklearn methods that
involve cloning in conjunction with accelerate in a multi-gpu setting.
This was due to sklearn creating copies of the Accelerator object. The
skorch docs proposed a fix, but this fix is now released in accelerate
itself. Therefore, the proposal can be safely removed.

Another minor change was to adjust the criterion in an example script.
This is not essential for the script to work, but prevents nan losses.