Better error message when missing classes_ attr
On NeuralNetClassifier, when classes_ is accessed but not known, we try
to use classes_inferred_ instead. If that is also not defined, users
currently just get an AttributeError that classes_inferred_ does not
exist. This can be very confusing.

This PR intercepts the error, provides a (hopefully) helpful error
message, then reraises. It is unfortunately not easily possible to
deduce what exactly went wrong to get to this, thus the error message
makes an educated guess.